Moved blackboard_item.dart
