+++ app/lib/blackboard/blackboard_dialog.dart
@@ -13,6 +13,7 @@ import 'package:build_context/build_context.dart';
 import 'package:filesharing_logic/filesharing_logic_models.dart';
 import 'package:flutter/material.dart';
 import 'package:group_domain_models/group_domain_models.dart';
+import 'package:sharezone/blackboard/blackboard_item.dart';
 import 'package:sharezone/blackboard/blackboard_page.dart';
 import 'package:sharezone/blackboard/blackboard_picture.dart';
 import 'package:sharezone/blackboard/blocs/blackboard_dialog_bloc.dart';
@@ -21,7 +22,6 @@ import 'package:sharezone/filesharing/dialog/attach_file.dart';
 import 'package:sharezone/filesharing/dialog/course_tile.dart';
 import 'package:sharezone/markdown/markdown_analytics.dart';
 import 'package:sharezone/markdown/markdown_support.dart';
-import 'package:sharezone/models/blackboard_item.dart';
 import 'package:sharezone/widgets/blackboard/blackboard_card.dart';
 import 'package:sharezone/widgets/material/list_tile_with_description.dart';
 import 'package:sharezone_common/api_errors.dart';
