+++ app/lib/blackboard/blackboard_item.dart
