+++ app/lib/blackboard/blocs/blackboard_dialog_bloc.dart
@@ -11,8 +11,8 @@ import 'package:files_basics/local_file.dart';
 import 'package:filesharing_logic/filesharing_logic_models.dart';
 import 'package:group_domain_models/group_domain_models.dart';
 import 'package:rxdart/rxdart.dart';
+import 'package:sharezone/blackboard/blackboard_item.dart';
 import 'package:sharezone/markdown/markdown_analytics.dart';
-import 'package:sharezone/models/blackboard_item.dart';
 import 'package:sharezone/util/API.dart';
 import 'package:sharezone_common/api_errors.dart';
 import 'package:sharezone_common/blackboard_validators.dart';
