+++ app/lib/dashboard/bloc/dashboard_bloc.dart
@@ -14,10 +14,10 @@ import 'package:design/design.dart';
 import 'package:firebase_hausaufgabenheft_logik/firebase_hausaufgabenheft_logik.dart';
 import 'package:group_domain_models/group_domain_models.dart';
 import 'package:rxdart/rxdart.dart';
+import 'package:sharezone/blackboard/blackboard_item.dart';
 import 'package:sharezone/calendrical_events/models/calendrical_event.dart';
 import 'package:sharezone/dashboard/models/homework_view.dart';
 import 'package:sharezone/dashboard/timetable/lesson_view.dart';
-import 'package:sharezone/models/blackboard_item.dart';
 import 'package:sharezone/timetable/src/bloc/timetable_bloc.dart';
 import 'package:sharezone/timetable/src/models/lesson.dart';
 import 'package:sharezone/timetable/src/models/lesson_data_snapshot.dart';
